fix "unsupported operation" error on set command (#110)
Set command outputs an error "unsupported operation". This error is
caused by #96, which removed cached settings. An user can update
properties by "set" command and the updated values are stored in the
cached settings. This change reverts caches settings and names it
"TransientSettings". The loaded settings and updated properties by the
user are stored in it.