feat: support regexp and ignorecase on find (#349)
The regular expression (RegExp) and ignore case search are now supported
on find mode. The RegExp search is based on the JavaScript [RegExp][]
syntax. This change adds two new properties to configure the search
mode:

- `findmode`: to switch between the normal and RegExp search.
- `ignorecase`: to enable or disable the ignore case search.

To enable the RegExp search, set the `findmode` property to `regexp`.

```
:set findmode=regexp
```

To enable the ignore case search, set the boolean `ignorecase` property:

```
:set ignorecase
```

[RegExp]:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/RegExp